backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openpyxl==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
reportlab==4.0.7
qrcode[pil]==7.4.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
